final commit KirkmanSolver.java
prints the first day of girls walking home successfully. The rest of the problem remains unfinished... This file has two different codes in it, it has my attempt at writing my own code for the Kirkman problem and it has the code I wrote based off of another user's code, this is all explained in the comments throughout this file.